Moved graphql_content_mutation to graphql_core. Deprecated graphql_entity_reference.
