+++ modules/graphql_core/tests/queries/delete.gql
