+++ modules/graphql_core/tests/src/Kernel/Mutations/UpdateEntityTest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Drupal\Tests\graphql_content_mutation\Kernel;
+namespace Drupal\Tests\graphql_core\Kernel\Mutations;
 
 use Drupal\Core\Entity\Entity\EntityViewDisplay;
 use Drupal\Core\Entity\Entity\EntityViewMode;
@@ -27,9 +27,6 @@ class UpdateEntityTest extends GraphQLFileTestBase {
     'text',
     'filter',
     'graphql_core',
-    'graphql_content',
-    'graphql_mutation',
-    'graphql_content_mutation',
   ];
 
   /**
@@ -47,17 +44,6 @@ class UpdateEntityTest extends GraphQLFileTestBase {
       ->grantPermission('access content')
       ->grantPermission('edit any test content')
       ->save();
-
-    $this->container->get('config.factory')->getEditable('graphql_content_mutation.schema')
-      ->set('types', [
-        'node' => [
-          'bundles' => [
-            'test' => [
-              'update' => TRUE,
-            ],
-          ],
-        ],
-      ])->save();
   }
 
   /**
