+++ modules/graphql_entity_reference/graphql_entity_reference.info.yml
@@ -3,5 +3,6 @@ type: module
 description: 'Traverse entity references in GraphQL queries.'
 package: GraphQL
 core: 8.x
+hidden: true
 dependencies:
   - graphql_content
